+++ @ornikar/eslint-config-formatjs/plugins/formatjs.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+  plugins: ['formatjs'],
+  rules: {
+    'formatjs/enforce-placeholders': 'error',
+    'formatjs/no-emoji': 'error',
+  },
+};