+++ @ornikar/prettier-config/CHANGELOG.md
@@ -3,6 +3,72 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0](https://github.com/ornikar/shared-configs/compare/@ornikar/prettier-config@2.0.0...@ornikar/prettier-config@3.0.0) (2020-04-21)
+
+
+### Features
+
+* require prettier version to new major [no issue] ([#353](https://github.com/ornikar/shared-configs/issues/353)) ([e3e8237](https://github.com/ornikar/shared-configs/commit/e3e8237))
+* update node version to 12 [no issue] ([#354](https://github.com/ornikar/shared-configs/issues/354)) ([6276917](https://github.com/ornikar/shared-configs/commit/6276917))
+
+
+### BREAKING CHANGES
+
+* prettier version
+
+* Update @ornikar/eslint-config/package.json
+
+Co-Authored-By: Christophe Hurpeau <christophe@hurpeau.com>
+
+* Update @ornikar/stylelint-config/package.json
+
+Co-Authored-By: Christophe Hurpeau <christophe@hurpeau.com>
+
+* Update @ornikar/prettier-config/package.json
+
+Co-Authored-By: Christophe Hurpeau <christophe@hurpeau.com>
+
+* Update package.json
+
+Co-authored-by: Christophe Hurpeau <christophe@hurpeau.com>
+* node version
+
+* Update config.yml
+
+
+
+
+
+# [2.0.0](https://github.com/ornikar/shared-configs/compare/@ornikar/prettier-config@1.0.1...@ornikar/prettier-config@2.0.0) (2019-12-13)
+
+
+### Features
+
+* eslint 6 [no issue] ([#298](https://github.com/ornikar/shared-configs/issues/298)) ([4bd9bf8](https://github.com/ornikar/shared-configs/commit/4bd9bf8))
+
+
+### BREAKING CHANGES
+
+* changed peer deps eslint 6, prettier, and enabled new rules
+
+* refactor: comment on liwi
+
+* fix: remove todo rule, now allowed with unicorn
+
+* chore: update @typescript-eslint
+
+* feat: explicit-function-return-type allow expressions and hof
+
+* feat: remove allowExpressions: true
+
+* feat: remove allowExpressions: true
+
+* fix: prefer flat map
+
+
+
+
+
 ## [1.0.1](https://github.com/ornikar/shared-configs/compare/@ornikar/prettier-config@1.0.0...@ornikar/prettier-config@1.0.1) (2019-11-27)
 
 **Note:** Version bump only for package @ornikar/prettier-config