+++ @ornikar/renovate-config/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.13.3...@ornikar/renovate-config@2.0.0) (2020-04-21)
+
+
+### Features
+
+* update node version to 12 [no issue] ([#354](https://github.com/ornikar/shared-configs/issues/354)) ([6276917](https://github.com/ornikar/shared-configs/commit/6276917))
+
+
+### BREAKING CHANGES
+
+* node version
+
+* Update config.yml
+
+
+
+
+
+## [1.13.3](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.13.2...@ornikar/renovate-config@1.13.3) (2020-01-22)
+
+**Note:** Version bump only for package @ornikar/renovate-config
+
+
+
+
+
 ## [1.13.2](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.13.1...@ornikar/renovate-config@1.13.2) (2019-10-09)